app_admincli/requirements.txt
bcrypt>=4.0
cryptography>=41.0
python-dotenv>=1.0
supabase>=2.0
Pillow>=10.0
requests>=2.31
mammoth>=1.6
pypdf>=5.0
